Add control clear
Adds an additional configuration parameter, controlkeyclear. When set to 1,
slock will no longer change to the failure color if a control key is pressed
while the buffer is empty. This is useful if, for example, you wake your
monitor up by pressing a control key and don't want to spoil the detection of
failed unlocking attempts.

https://tools.suckless.org/slock/patches/control-clear/slock-git-20161012-control-clear.diff