Fix wide glyphs truncation
ST divides rendered text to sequences of glyphs with same style and then sequentially draws them.
This leads to a problem: if a wide glyph is followed by another glyph with different style
(can be color, weight or anything), the excessive parts of the glyph will be overdrawn by the
beginning of the next sequence.

- Use 2 passes for drawing lines: one for background, another one for foreground.
  So parts of wide glyphs do not get overdrawn.
- Git rid of rectangle clipping. It was introduced in this db6f796 and is intended for
  removing overdrawing for some fonts.
  Such fonts are still going to look crippled and are not worth using in the first place.