Add font2
This patch allows to add spare font besides default. Some glyphs can be not present in default
font. For this glyphs st uses font-config and try to find them in font cache first. This patch
append fonts defined in font2 variable to the beginning of font cache. So they will be used first
for glyphs that absent in default font.

https://st.suckless.org/patches/font2/st-font2-0.8.5.diff